+++ R/get_Nalls2019_loci.R
@@ -11,6 +11,8 @@
 #'  (default: \code{TRUE}), or the data itself as a named list (\code{FALSE}).
 #' @param return_dir Return the directory name
 #'  instead of the individual file paths. 
+#' @param limit_snps Limit the number of SNPs saved in each file.
+#' @param force_new Force the creation of new files even if old ones exists.
 #' @param verbose Print messages. 
 #' @inheritParams get_data
 #' 
@@ -19,10 +21,11 @@
 #' @returns File paths 
 #' @examples 
 #' files <- get_Nalls2019_loci()
-get_Nalls2019_loci <- function(save_dir=tools::R_user_dir(package = "echodata",
-                                                          which = "cache"),
+get_Nalls2019_loci <- function(save_dir=tempdir(),
                                return_paths=TRUE,
                                return_dir=FALSE,
+                               limit_snps=NULL,
+                               force_new=FALSE,
                                verbose=TRUE){
     
     dataset <- file.path(save_dir,"Nalls23andMe_2019")
@@ -30,6 +33,9 @@ get_Nalls2019_loci <- function(save_dir=tools::R_user_dir(package = "echodata",
     loci <- list("BST1"=echodata::BST1,
                  "LRRK2"=echodata::LRRK2,
                  "MEX3C"=echodata::MEX3C)
+    if(!is.null(limit_snps)){
+        loci <- lapply(loci, function(x)x[seq_len(limit_snps)])
+    }
     #### Return early ####
     if(isFALSE(return_paths)) return(loci)
     #### Write to disk ####
@@ -39,7 +45,8 @@ get_Nalls2019_loci <- function(save_dir=tools::R_user_dir(package = "echodata",
     )
     names(files) <- names(loci)
     #### Return early if files already written ####
-    if(all(file.exists(unlist(files)))) {
+    if(all(file.exists(unlist(files))) && 
+       isFALSE(force_new)) {
         if(isTRUE(return_dir)) files <- dirname(files[[1]])
         return(files)
     }