+++ NEWS.md
@@ -12,7 +12,10 @@
     - `results_report`
     - `merge_finemapping_results`
 * `portal_query`: add arg to return results in table format.
-* Add and export `set_permissions` and `get_os` functions. 
+* Add and export new functions:
+    - `set_permissions`
+    - `get_os`
+    - `find_executable`
 
 ## Bug fixes